% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_abide.R
\name{aparcstats_files_ABIDE}
\alias{aparcstats_files_ABIDE}
\title{Return a hemilist with the paths of the aparc stats files for the ABIDE dataset.}
\usage{
aparcstats_files_ABIDE()
}
\value{
a named list with keys 'lh' and 'rh', both hold a filepath to the parcellation statistics file for the respective hemisphere.
}
\description{
Return a hemilist with the paths of the aparc stats files for the ABIDE dataset. The files come pre-computed with the brainnet package. They were created by running the FreeSurfer tool 'aparcstats2table' on the full ABIDE I dataset (subjects with structural image data only, of course), with cortical thickness as the 'measure' argument.
}
\seealso{
Other ABIDE functions: 
\code{\link{load_ABIDE_metadata_males}()},
\code{\link{load_ABIDE_metadata}()}
}
\concept{ABIDE functions}
