% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_ixi.R
\name{aparcstats_files_IXI}
\alias{aparcstats_files_IXI}
\title{Return a hemilist with the paths of the aparc stats files for the IXI dataset.}
\usage{
aparcstats_files_IXI(measure = "thickness")
}
\arguments{
\item{measure}{character string, a measure for which the stats files are available. Typically one of 'thickness', 'volume', 'area', unless you have computed more and added them manually.}
}
\value{
a named list with keys 'lh' and 'rh', both hold a filepath to the parcellation statistics file for the respective hemisphere.
}
\description{
Return a hemilist with the paths of the aparc stats files for the IXI dataset. The files come pre-computed with the brainnet package. They were created by running the FreeSurfer tool 'aparcstats2table' on the full IXI dataset (subjects with structural image data only, of course) several times (with thickness, volume and area as the 'measure' argument).
}
\note{
The asegstats (for the volume) are also available in 'inst/extdata' folder of this package as the file 'IXI_asegstats.tsv', but the path cannot be retrieved with this function as there are no hemi-specific files, just get it manually.
}
\seealso{
Other ABIDE functions: 
\code{\link{aparcstats_files_ABIDE}()},
\code{\link{load_ABIDE_metadata_males}()},
\code{\link{load_ABIDE_metadata}()}
}
\concept{ABIDE functions}
