% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{convert.segstats.table.format}
\alias{convert.segstats.table.format}
\title{Convert the parcstats data.frame to match the format used by fsbrain::group.agg.atlas.native.}
\usage{
convert.segstats.table.format(dt)
}
\arguments{
\item{dt}{data.frame, as returned by \code{read.segstats.table}}
}
\value{
converted data.frame
}
\description{
Convert the parcstats data.frame to match the format used by fsbrain::group.agg.atlas.native.
}
\keyword{internal}
