% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics_plots.R
\name{df_group_plot}
\alias{df_group_plot}
\title{Plot all columns of a data.frame by group.}
\usage{
df_group_plot(dtf, group_column = "group")
}
\arguments{
\item{df}{data.frame, the input data. Can contain factor columns and continuous columns.}

\item{group_by}{character string, the column name that we should group by.}
}
\description{
Creates a plot with one subplot per data.frame column. Continuous data is plotted as histograms, factors as bar plots.
}
\examples{
\dontrun{
md = load_ABIDE_metadata_males();
dtf = md$demographics;
df_group_plot(dtf);
}

}
\keyword{internal}
