% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin_plots.R
\name{effect_size_violin_plots}
\alias{effect_size_violin_plots}
\title{Create effect size violin plots for all predictors at once.}
\usage{
effect_size_violin_plots(effects_mat)
}
\arguments{
\item{effects_mat}{n*m matrix in which the n rownames are the predictors (like age, sex, site, ...) and the m columns contain the effect sizes for the m vertices.}
}
\value{
a ggplot instance.
}
\description{
Create effect size violin plots for all predictors at once.
}
