% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Compute and print metrics for classification model evaluation}
\usage{
evaluate_model(actual, predicted)
}
\arguments{
\item{actual}{vector of actual labels}

\item{predicted}{vector of predicted labels}
}
\value{
Nothing, called for printing side effect.
}
\description{
Compute and print metrics for classification model evaluation
}
\author{
T Schaefer
}
