% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapath.R
\name{get_ABIDE_path_on_tims_machines}
\alias{get_ABIDE_path_on_tims_machines}
\title{Search for the ABIDE data directory on my machine.}
\usage{
get_ABIDE_path_on_tims_machines()
}
\value{
character string, the path to the ABIDE data, preprocessed with FreeSurfer (the recon-all output dir or SUBJECTS_DIR).
}
\description{
Search for the ABIDE data directory on my machine.
}
\note{
Ignore this internal function. You should simply hard-code the path to the data on your machine in your script. This function only exists to make the demo script short (and still allow it to run on all my machines during development).
}
\keyword{internal}
