% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_ixi.R
\name{get_IXI_brainstats_for_subjects}
\alias{get_IXI_brainstats_for_subjects}
\title{Load IXI brain stats and filter by subjects_list.}
\usage{
get_IXI_brainstats_for_subjects(subjects_list)
}
\arguments{
\item{subjects_list}{vector of character strings, the subjects to filter. Pass NULL to get full brain stats.}
}
\value{
data.frame with brain stats
}
\description{
Load IXI brain stats and filter by subjects_list.
}
\seealso{
Other IXI functions: 
\code{\link{load_IXI_metadata}()},
\code{\link{postproc_IXI_demographics}()}
}
\concept{IXI functions}
\keyword{internal}
