% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_abide.R
\name{load_ABIDE_metadata}
\alias{load_ABIDE_metadata}
\title{Load ABIDE demographics and brain stats.}
\usage{
load_ABIDE_metadata(impute_data = TRUE, add_merged = TRUE)
}
\arguments{
\item{impute_data}{logical, whether to impute missing data with mice. The filtered data (in_DiMartino_study AND has_structural_img) contains 30 subjects with missing IQ values.}

\item{add_merged}{logical, whether to add return a single merged data.frame to the return value list with key 'merged'.}
}
\value{
named list with entries 'brainstats', 'demographics' and 'subjects_list'. The first two are data.frames, the last one is a vector of character strings.
}
\description{
Load ABIDE demographics and brain stats and filter so that only subjects from the DiMartino study which have structural imaging data remain.
}
