% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_abide.R
\name{load_ABIDE_metadata_males}
\alias{load_ABIDE_metadata_males}
\title{Load ABIDE demographics and brain stats for male sujects from diMartino study.}
\usage{
load_ABIDE_metadata_males(
  impute_data = TRUE,
  exclude_bad_quality = c("UM_1_0050272")
)
}
\arguments{
\item{impute_data}{logical, whether to impute missing data with mice. The filtered data (in_DiMartino_study AND has_structural_img) contains 30 male subjects with missing IQ values.}

\item{exclude_bad_quality}{vector of character strings, the manually excluded subjects, e.g. due to bad quality}
}
\value{
named list with entries 'brainstats', 'demographics' and 'subjects_list'. The first two are data.frames, the last one is a vector of character strings.
}
\description{
Load ABIDE demographics and brain stats and filter so that only subjects from the DiMartino study \code{(https://www.nature.com/articles/mp201378)} which have structural imaging data remain. Note that the diMartino study only included male subjects.
}
\seealso{
Other ABIDE functions: 
\code{\link{aparcstats_files_ABIDE}()},
\code{\link{load_ABIDE_metadata}()}
}
\concept{ABIDE functions}
