% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_brainmap.R
\name{load_brainmap_metadata}
\alias{load_brainmap_metadata}
\title{Load Brainmap demographics and brain stats.}
\usage{
load_brainmap_metadata()
}
\value{
named list with entries 'brainstats', 'demographics' and 'subjects_list'. The first two are data.frames, the last one is a vector of character strings.
}
\description{
Load Brainmap demographics and brain stats.
}
