% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_ixi.R
\name{postproc_IXI_demographics}
\alias{postproc_IXI_demographics}
\title{Post-process IXI demographiocs data}
\usage{
postproc_IXI_demographics(demographics)
}
\arguments{
\item{demographics}{the IXI demographics data.frame}
}
\value{
fixed df
}
\description{
Post-process IXI demographiocs data
}
\seealso{
Other IXI functions: 
\code{\link{get_IXI_brainstats_for_subjects}()},
\code{\link{load_IXI_metadata}()}
}
\author{
T Schaefer
}
\concept{IXI functions}
