% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read.segstats.table}
\alias{read.segstats.table}
\title{Read aparcstats or segstats table file created by FreeSurfer tools.}
\usage{
read.segstats.table(filepath)
}
\arguments{
\item{filepath}{character string, the path to the file.}
}
\value{
data.frame with seg/parc stats
}
\description{
Read aparcstats or segstats table file created by FreeSurfer tools.
}
