% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_segstats_table}
\alias{read_segstats_table}
\title{Read aparcstats or segstats table file created by FreeSurfer tool 'aparcstats2table' or 'asegstats2table'.}
\usage{
read_segstats_table(filepath)
}
\arguments{
\item{filepath}{character string, the path to the aparcstats or asegstats file.}
}
\value{
data.frame with seg/parc stats
}
\description{
Read aparcstats or segstats table file created by FreeSurfer tool 'aparcstats2table' or 'asegstats2table'.
}
\examples{
\dontrun{
stats_file ="~/study1/lh.aparc_allsubjects.table";
df = read_segstats_table(stats_file);
}

}
\author{
T Schaefer
}
