% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{region.data.from.segstat.tables}
\alias{region.data.from.segstat.tables}
\title{Retrieve region-based morphometry data from a parcellation stats files created by 'aparcstats2table'.}
\usage{
region.data.from.segstat.tables(lh_segstatsfile, rh_segstatsfile)
}
\arguments{
\item{lh_segstatsfile}{character string, the path to the aparcstats file for the left hemisphere.}

\item{rh_segstatsfile}{character string, the path to the aparcstats file for the right hemisphere.}
}
\value{
data.frame, see \code{\link[fsbrain]{group.agg.atlas.native}} for details as this one is designed to match that format.
}
\description{
Retrieve region-based morphometry data from a parcellation stats files created by 'aparcstats2table'.
}
\note{
The parcellation stats files must be generated manually for your sample using 'aparcstats2table'.
}
