% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{region_data_from_segstat_tables}
\alias{region_data_from_segstat_tables}
\title{Retrieve region-based morphometry data from a parcellation stats files created by 'aparcstats2table'.}
\usage{
region_data_from_segstat_tables(lh_segstatsfile, rh_segstatsfile)
}
\arguments{
\item{lh_segstatsfile}{character string, the path to the aparcstats file for the left hemisphere.}

\item{rh_segstatsfile}{character string, the path to the aparcstats file for the right hemisphere.}
}
\value{
data.frame, see \code{\link[fsbrain]{group.agg.atlas.native}} for details as this one is designed to match that format.
}
\description{
Retrieve region-based morphometry data from a parcellation stats files created by 'aparcstats2table'.
}
\note{
The parcellation stats files must be generated manually for your sample using 'aparcstats2table'.
}
\examples{
\dontrun{
lh_sfile ="~/study1/lh.aparc_allsubjects.table";
rh_sfile ="~/study1/rh.aparc_allsubjects.table";
df = region_data_from_segstat_tables(lh_sfile, rr_sfile);
}

}
\author{
T Schaefer
}
