% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_glm.R
\name{slm_effect_sizes}
\alias{slm_effect_sizes}
\title{Compute effect sizes for mass-univariate GLM analysis.}
\usage{
slm_effect_sizes(
  X,
  Y,
  predictors,
  output = c("F", "p", "etasq", "partial.etasq", "cohens.f", "rsq", "power")
)
}
\arguments{
\item{X}{numerical matrix, the design or model matrix, typically created from the demographics data using \code{stats::model.matrix}.}

\item{Y}{numerical matrix, the target value, typically neuroimaging data}

\item{predictors}{vector of character strings, the names of the predictors in the model matrix}

\item{output}{vector of pre-defined character strings, defined what values to return. Leave alone if in doubt.}
}
\value{
see output parameter
}
\description{
Compute effect sizes for mass-univariate GLM analysis.
}
\author{
C Ecker
}
