% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_glm.R
\name{slm_t}
\alias{slm_t}
\title{Compute t map for mass-univariate GLM analysis.}
\usage{
slm_t(X, Y, model.term, output = c("t", "p", "p.adjust"))
}
\description{
Compute t map for mass-univariate GLM analysis.
}
\author{
C Ecker
}
\keyword{internal}
