% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_glm.R
\name{slm_threshold_statistical_map}
\alias{slm_threshold_statistical_map}
\title{Threshold a statistical map.}
\usage{
slm_threshold_statistical_map(x, p, alpha, p.adjust.method = "none")
}
\description{
Threshold a statistical map.
}
\author{
C Ecker
}
\keyword{internal}
