% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_ixi.R
\name{load_IXI_metadata}
\alias{load_IXI_metadata}
\title{Load IXI demographics and brain stats}
\usage{
load_IXI_metadata()
}
\value{
named list with entries 'brainstats', 'demographics' and 'subjects_list'. The first two are data.frames, the last one is a vector of character strings.
}
\description{
Load IXI demographics and brain stats
}
\note{
The column that you would expect to be named 'subject_id' is called 'subject_data_dirname' in demographics/merged, because the IXI ID is not the same as the FreeSurfer subject directory name here.
}
\seealso{
Other IXI functions: 
\code{\link{get_IXI_brainstats_for_subjects}()},
\code{\link{postproc_IXI_demographics}()}
}
\concept{IXI functions}
