% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_glm.R
\name{slm_effect_sizes}
\alias{slm_effect_sizes}
\title{Compute effect sizes for mass-univariate GLM analysis.}
\usage{
slm_effect_sizes(
  X,
  Y,
  predictors,
  output = c("F", "p", "etasq", "partial.etasq", "cohens.f", "rsq", "power")
)
}
\description{
Compute effect sizes for mass-univariate GLM analysis.
}
\author{
C Ecker
}
\keyword{internal}
