% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_glm.R
\name{slm_F}
\alias{slm_F}
\title{Compute F stats map for mass-univariate GLM analysis.}
\usage{
slm_F(X, Y, predictors, output = c("F", "p", "p.adjust"))
}
\description{
Compute F stats map for mass-univariate GLM analysis.
}
\author{
C Ecker
}
\keyword{internal}
