% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_glm.R
\name{slm_threshold_statistical_map}
\alias{slm_threshold_statistical_map}
\title{Threshold a statistical map by setting all values with p > alpha to NaN.}
\usage{
slm_threshold_statistical_map(x, p, alpha = 0.05, p.adjust.method = "none")
}
\arguments{
\item{x}{vector of input data values (not p values, see parameter p for those). The ones not surviving the correction for multiple comparisons of their respective p value will be set to NaN in the output.}

\item{p}{the p-values for the x values}

\item{alpha}{the alpha level}

\item{p.adjust.method}{passed on to \code{stats::p.adjust}}
}
\value{
a version of x in which the values that did not survive the correction for multiple comparisons of their respective p value are set to NaN.
}
\description{
Threshold a statistical map by setting all values with p > alpha to NaN.
}
\examples{
    braindata = rnorm(40, 5.0, 0.2);
    braindata_p = rnorm(40, 0.05, 0.1);
    data_thresh = slm_threshold_statistical_map(braindata, braindata_p);

}
\author{
C Ecker, documentation by T Schaefer
}
